Updated module files, added env var pathing, fixed other minor issues
