+++ cipher_cando/triggers/trigger.py
@@ -4,7 +4,7 @@ import sys
 import argparse
 
 sys.path.append("../../")
-from module_properties.utils.properties import insert_properties_from_smiles
+from cipher_identifiers.utils.compounds import id_compound_from_smiles
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--testing", action="store_true")
@@ -32,28 +32,23 @@ else:
         + "@aspirecluster0.hmj3q.mongodb.net/cipher_aspire?retryWrites=true&w=majority"
     )
 
+
 me.connect(host=URI)
 MONGO_CLIENT = pymongo.MongoClient(URI)
 
 
-def properties_trigger():
+def cando_trigger():
     if args.testing:
         compounds_coll = MONGO_CLIENT.cipher_testing.compounds
     else:
         compounds_coll = MONGO_CLIENT.cipher_aspire.compounds
-
     try:
         with compounds_coll.watch([{"$match": {"operationType": "insert"}}]) as stream:
             for change in stream:
-                doc = change["fullDocument"]
-                smiles = doc["smiles"]
-                try:
-                    insert_properties_from_smiles(smiles)
-                except Exception as e:
-                    print(e)
+                pass
     except Exception as e:
         print(e)
 
 
 if __name__ == "__main__":
-    properties_trigger()
+    cando_trigger()
