+++ cipher_identifiers/docs/docs.py
@@ -11,9 +11,8 @@ def validate_smiles(val):
 
 
 class Compounds(me.Document):
+    inchikey = me.StringField(required=True, primary_key=True)
     name = me.StringField(default="")
-    id = me.StringField(required=True, primary_key=True)
-    inchikey = me.StringField(required=True)
     smiles = me.StringField(required=True, validation=validate_smiles)
     inchi = me.StringField(required=True)
     cid = me.IntField(default=-1)
@@ -28,7 +27,7 @@ def check_inchikey_in_compounds(val):
 
 class Substances(me.Document):
     # Cipher SID
-    id = me.StringField(required=True, primary_key=True)
+    cipher_sid = me.StringField(required=True, primary_key=True)
     smiles = me.StringField(required=True, validation=validate_smiles)
     inchikey = me.StringField(required=True, validation=check_inchikey_in_compounds)
     source = me.StringField(required=True)
@@ -38,7 +37,7 @@ class Substances(me.Document):
 
 class Models(me.Document):
     # Cipher MID
-    id = me.StringField(required=True, primary_key=True)
+    cipher_mid = me.StringField(required=True, primary_key=True)
     source = me.StringField(required=True)
     parameters = me.StringField(default="\{\}")
     modified = me.DateTimeField(default=datetime.datetime.utcnow)
@@ -51,7 +50,7 @@ def check_mid_in_models(val):
 
 class External_Databases(me.Document):
     # Cipher EXDBID
-    id = me.StringField(required=True, primary_key=True)
+    cipher_exdbid = me.StringField(required=True, primary_key=True)
     source = me.StringField(required=True)
     url = me.StringField(default="")
     modified = me.DateTimeField(default=datetime.datetime.utcnow)
@@ -59,7 +58,7 @@ class External_Databases(me.Document):
 
 class Biomolecules(me.Document):
     # Cipher BMID
-    id = me.StringField(required=True, primary_key=True)
+    ciper_bmid = me.StringField(required=True, primary_key=True)
     name = me.StringField(default="")
     pdb_id = me.StringField(default="")
     chain_id = me.StringField(default="")
@@ -77,7 +76,7 @@ def check_bmid_in_biomolecules(val):
 
 class Binding_Sites(me.Document):
     # Chiper BSID
-    id = me.StringField(required=True, primary_key=True)
+    cipher_bsid = me.StringField(required=True, primary_key=True)
     cipher_bmid = me.StringField(required=True, validation=check_bmid_in_biomolecules)
     template_pdb_id = me.StringField(default="")
     template_chain_id = me.StringField(default="")
@@ -94,7 +93,7 @@ def check_bsid_in_binding_sites(val):
 
 class Ligands(me.Document):
     # Cipher LID
-    id = me.StringField(required=True, primary_key=True)
+    cipher_lid = me.StringField(required=True, primary_key=True)
     inchikey = me.StringField(required=True, validation=check_inchikey_in_compounds)
     pdb_ligand_id = me.StringField(default="")
     # TODO: check this to ensure that it is in the database
