+++ cipher_identifiers/triggers/trigger.py
@@ -1,4 +1,4 @@
-from os import remove
+import os
 import mongoengine as me
 import pymongo
 import sys
@@ -6,34 +6,16 @@ import argparse
 from rdkit import Chem
 
 sys.path.append("../../")
-from module_identifiers.utils.compounds import id_compound_from_smiles
+from cipher_identifiers.utils.compounds import id_compound_from_smiles
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--testing", action="store_true")
 args = parser.parse_args()
 
-LOGIN = open("../../utils/login.txt", "r")
-USERNAME = LOGIN.readline().replace("\n", "")
-PASSWORD = LOGIN.readline().replace("\n", "")
-LOGIN.close()
-
 if args.testing:
-    URI = (
-        "mongodb+srv://"
-        + USERNAME
-        + ":"
-        + PASSWORD
-        + "@aspirecluster0.hmj3q.mongodb.net/cipher_testing?retryWrites=true&w=majority"
-    )
+    URI = os.environ["TESTING_URI"]
 else:
-    URI = (
-        "mongodb+srv://"
-        + USERNAME
-        + ":"
-        + PASSWORD
-        + "@aspirecluster0.hmj3q.mongodb.net/cipher_aspire?retryWrites=true&w=majority"
-    )
-
+    URI = os.environ["MONGO_URI"]
 
 me.connect(host=URI)
 MONGO_CLIENT = pymongo.MongoClient(URI)
