+++ cipher_properties/triggers/trigger.py
@@ -1,10 +1,11 @@
+import os
 import mongoengine as me
 import pymongo
 import sys
 import argparse
 
 sys.path.append("../../")
-from module_identifiers.utils.compounds import id_compound_from_smiles
+from cipher_properties.utils.properties import insert_properties_from_smiles
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--testing", action="store_true")
@@ -16,39 +17,32 @@ PASSWORD = LOGIN.readline().replace("\n", "")
 LOGIN.close()
 
 if args.testing:
-    URI = (
-        "mongodb+srv://"
-        + USERNAME
-        + ":"
-        + PASSWORD
-        + "@aspirecluster0.hmj3q.mongodb.net/cipher_testing?retryWrites=true&w=majority"
-    )
+    URI = os.environ["TESTING_URI"]
 else:
-    URI = (
-        "mongodb+srv://"
-        + USERNAME
-        + ":"
-        + PASSWORD
-        + "@aspirecluster0.hmj3q.mongodb.net/cipher_aspire?retryWrites=true&w=majority"
-    )
-
+    URI = os.environ["MONGO_URI"]
 
 me.connect(host=URI)
 MONGO_CLIENT = pymongo.MongoClient(URI)
 
 
-def cando_trigger():
+def properties_trigger():
     if args.testing:
         compounds_coll = MONGO_CLIENT.cipher_testing.compounds
     else:
         compounds_coll = MONGO_CLIENT.cipher_aspire.compounds
+
     try:
         with compounds_coll.watch([{"$match": {"operationType": "insert"}}]) as stream:
             for change in stream:
-                pass
+                doc = change["fullDocument"]
+                smiles = doc["smiles"]
+                try:
+                    insert_properties_from_smiles(smiles)
+                except Exception as e:
+                    print(e)
     except Exception as e:
         print(e)
 
 
 if __name__ == "__main__":
-    cando_trigger()
+    properties_trigger()
