+++ cipher_properties/utils/old.py
